+++ crates/runner/mitm-addon/tests/test_anthropic_messages.py
@@ -1,9 +1,14 @@
 """Tests for Anthropic Messages usage extraction."""
 
+import gzip
+import json
+
 import pytest
 
 from usage import (
     create_anthropic_messages_sse_usage_extractor,
+    extract_anthropic_messages_usage_from_json,
+    extract_anthropic_messages_usage_with_error_from_json,
 )
 
 
@@ -449,3 +454,104 @@ class TestAnthropicSseUsageExtractor:
         assert usage["tokens.output"] == 300
         assert usage["tokens.cache_read"] == 6000  # updated
         assert usage["tokens.input"] == 100  # unchanged (not in delta)
+
+
+class TestExtractAnthropicUsageFromJson:
+    """Tests for extract_anthropic_messages_usage_from_json helper."""
+
+    def test_extracts_model_and_tokens(self):
+        body = b'{"model":"claude-sonnet-4-6","usage":{"input_tokens":100,"output_tokens":500}}'
+        result = extract_anthropic_messages_usage_from_json(body, None)
+        assert result == {
+            "model": "claude-sonnet-4-6",
+            "tokens.input": 100,
+            "tokens.output": 500,
+        }
+
+    def test_extracts_cache_tokens(self):
+        body = (
+            b'{"model":"claude-sonnet-4-6","usage":'
+            b'{"input_tokens":10,"output_tokens":5,'
+            b'"cache_read_input_tokens":50,"cache_creation_input_tokens":0}}'
+        )
+        result = extract_anthropic_messages_usage_from_json(body, None)
+        assert result is not None
+        assert result["tokens.cache_read"] == 50
+        assert result["tokens.cache_creation"] == 0
+
+    def test_gzip_compressed(self, headers):
+        original = b'{"model":"test","usage":{"input_tokens":42}}'
+        compressed = gzip.compress(original)
+        headers = headers(("Content-Encoding", "gzip"))
+        result = extract_anthropic_messages_usage_from_json(compressed, headers)
+        assert result is not None
+        assert result["model"] == "test"
+        assert result["tokens.input"] == 42
+
+    def test_truncated_gzip_stays_silent_but_diagnostic_returns_error(self, headers):
+        original = b'{"model":"test","usage":{"input_tokens":42}}'
+        truncated = gzip.compress(original)[:10]
+        headers = headers(("Content-Encoding", "gzip"))
+
+        assert extract_anthropic_messages_usage_from_json(truncated, headers) is None
+        usage, error = extract_anthropic_messages_usage_with_error_from_json(truncated, headers)
+        assert usage is None
+        assert error == "incomplete compressed body"
+
+    def test_invalid_json_returns_none(self):
+        assert extract_anthropic_messages_usage_from_json(b"not json", None) is None
+
+    def test_no_usage_field_returns_none(self):
+        assert extract_anthropic_messages_usage_from_json(b'{"id":"msg_1"}', None) is None
+
+    def test_non_dict_returns_none(self):
+        assert extract_anthropic_messages_usage_from_json(b"[1,2,3]", None) is None
+
+    def test_ignores_unmapped_web_search_requests(self):
+        body = (
+            b'{"model":"claude-sonnet-4-6","usage":'
+            b'{"input_tokens":10,"output_tokens":5,'
+            b'"server_tool_use":{"web_search_requests":2}}}'
+        )
+        result = extract_anthropic_messages_usage_from_json(body, None)
+        assert result is not None
+        assert "web_search_requests" not in result
+        assert result["tokens.input"] == 10
+
+    def test_ignores_invalid_usage_quantities(self):
+        body = (
+            b'{"model":"claude-sonnet-4-6","usage":'
+            b'{"input_tokens":-1,"output_tokens":5,'
+            b'"cache_read_input_tokens":"50",'
+            b'"cache_creation_input_tokens":true}}'
+        )
+        result = extract_anthropic_messages_usage_from_json(body, None)
+        assert result == {
+            "model": "claude-sonnet-4-6",
+            "tokens.output": 5,
+        }
+
+    def test_handles_large_gzipped_body(self, headers):
+        """Body that decompresses past the legacy 64 KB cap should still parse.
+
+        Regression test for the silent 64 KB default in body_decoding.decompress_body
+        which used to truncate large model-provider non-SSE responses and cause
+        usage extraction to silently fail.
+        """
+        # Raw body > 64 KB (legacy STREAM_BUFFER_LIMIT) so the bug, if reintroduced,
+        # would truncate decompression output and break json.loads below.
+        big_text = "x" * (100 * 1024)
+        payload = json.dumps(
+            {
+                "id": "msg_1",
+                "model": "claude-sonnet-4-6",
+                "content": [{"type": "text", "text": big_text}],
+                "usage": {"input_tokens": 50, "output_tokens": 100},
+            }
+        ).encode()
+        compressed = gzip.compress(payload)
+        headers = headers(("Content-Encoding", "gzip"))
+        result = extract_anthropic_messages_usage_from_json(compressed, headers)
+        assert result is not None
+        assert result["tokens.input"] == 50
+        assert result["tokens.output"] == 100
