+++ crates/runner/mitm-addon/tests/test_body_capture.py
@@ -2,7 +2,6 @@
 
 import base64
 import gzip
-import json
 import zlib
 
 import brotli
@@ -17,48 +16,8 @@ from body_capture import (
     _truncate_bytes_utf8_safe,
     add_capture_fields,
 )
-from body_decoding import (
-    create_stream_decode_feed,
-    decompress_body,
-)
-from body_limits import STREAM_BUFFER_LIMIT, STREAM_DECODE_CHUNK_LIMIT
-from usage import (
-    extract_anthropic_messages_usage_from_json,
-    extract_anthropic_messages_usage_with_error_from_json,
-    extract_openai_responses_usage_from_json,
-    extract_openai_responses_usage_with_error_from_json,
-)
-
-
-def _track_brotli_decompressor(monkeypatch):
-    real_decompressor = brotli.Decompressor
-    stats = {"calls": 0, "max_input": 0, "max_output": 0}
-
-    class CountingDecompressor:
-        def __init__(self):
-            self._inner = real_decompressor()
-
-        def process(self, chunk: bytes) -> bytes:
-            out = self._inner.process(chunk)
-            stats["calls"] += 1
-            stats["max_input"] = max(stats["max_input"], len(chunk))
-            stats["max_output"] = max(stats["max_output"], len(out))
-            return out
-
-        def is_finished(self) -> bool:
-            return self._inner.is_finished()
-
-    monkeypatch.setattr("body_decoding.brotli.Decompressor", CountingDecompressor)
-    return stats
-
-
-def _pseudo_random_ascii(size: int) -> bytes:
-    state = 0x12345678
-    body = bytearray()
-    for _ in range(size):
-        state = (1103515245 * state + 12345) & 0x7FFFFFFF
-        body.append(32 + (state % 95))
-    return bytes(body)
+from body_limits import STREAM_BUFFER_LIMIT
+from tests.body_decode_helpers import pseudo_random_ascii, track_brotli_decompressor
 
 
 class TestIsTextContent:
@@ -1248,13 +1207,13 @@ class TestDecompression:
         assert len(entry["response_body"]) == STREAM_BUFFER_LIMIT
 
     def test_brotli_large_text_uses_adaptive_chunks(self, real_flow, monkeypatch):
-        original = _pseudo_random_ascii(STREAM_BUFFER_LIMIT // 2)
+        original = pseudo_random_ascii(STREAM_BUFFER_LIMIT // 2)
         compressed = brotli.compress(original)
         old_call_count = (len(compressed) + 15) // 16
         assert len(compressed) < STREAM_BUFFER_LIMIT
         assert old_call_count > 1000
 
-        stats = _track_brotli_decompressor(monkeypatch)
+        stats = track_brotli_decompressor(monkeypatch)
 
         flow = self._make_flow_with_compressed_buffer(real_flow, compressed, "br", "text/plain")
         entry = {}
@@ -1271,7 +1230,7 @@ class TestDecompression:
         compressed = brotli.compress(original)
         assert len(compressed) < STREAM_BUFFER_LIMIT
 
-        stats = _track_brotli_decompressor(monkeypatch)
+        stats = track_brotli_decompressor(monkeypatch)
 
         flow = self._make_flow_with_compressed_buffer(real_flow, compressed, "br", "text/plain")
         entry = {}
@@ -1363,644 +1322,3 @@ class TestDecompression:
         entry = {}
         add_capture_fields(flow, entry)
         assert entry.get("response_body_truncated") is True or "response_body" not in entry
-
-
-class TestExtractAnthropicUsageFromJson:
-    """Tests for extract_anthropic_messages_usage_from_json helper."""
-
-    def test_extracts_model_and_tokens(self):
-        body = b'{"model":"claude-sonnet-4-6","usage":{"input_tokens":100,"output_tokens":500}}'
-        result = extract_anthropic_messages_usage_from_json(body, None)
-        assert result == {
-            "model": "claude-sonnet-4-6",
-            "tokens.input": 100,
-            "tokens.output": 500,
-        }
-
-    def test_extracts_cache_tokens(self):
-        body = (
-            b'{"model":"claude-sonnet-4-6","usage":'
-            b'{"input_tokens":10,"output_tokens":5,'
-            b'"cache_read_input_tokens":50,"cache_creation_input_tokens":0}}'
-        )
-        result = extract_anthropic_messages_usage_from_json(body, None)
-        assert result is not None
-        assert result["tokens.cache_read"] == 50
-        assert result["tokens.cache_creation"] == 0
-
-    def test_gzip_compressed(self, headers):
-        original = b'{"model":"test","usage":{"input_tokens":42}}'
-        compressed = gzip.compress(original)
-        headers = headers(("Content-Encoding", "gzip"))
-        result = extract_anthropic_messages_usage_from_json(compressed, headers)
-        assert result is not None
-        assert result["model"] == "test"
-        assert result["tokens.input"] == 42
-
-    def test_truncated_gzip_stays_silent_but_diagnostic_returns_error(self, headers):
-        original = b'{"model":"test","usage":{"input_tokens":42}}'
-        truncated = gzip.compress(original)[:10]
-        headers = headers(("Content-Encoding", "gzip"))
-
-        assert extract_anthropic_messages_usage_from_json(truncated, headers) is None
-        usage, error = extract_anthropic_messages_usage_with_error_from_json(truncated, headers)
-        assert usage is None
-        assert error == "incomplete compressed body"
-
-    def test_invalid_json_returns_none(self):
-        assert extract_anthropic_messages_usage_from_json(b"not json", None) is None
-
-    def test_no_usage_field_returns_none(self):
-        assert extract_anthropic_messages_usage_from_json(b'{"id":"msg_1"}', None) is None
-
-    def test_non_dict_returns_none(self):
-        assert extract_anthropic_messages_usage_from_json(b"[1,2,3]", None) is None
-
-    def test_ignores_unmapped_web_search_requests(self):
-        body = (
-            b'{"model":"claude-sonnet-4-6","usage":'
-            b'{"input_tokens":10,"output_tokens":5,'
-            b'"server_tool_use":{"web_search_requests":2}}}'
-        )
-        result = extract_anthropic_messages_usage_from_json(body, None)
-        assert result is not None
-        assert "web_search_requests" not in result
-        assert result["tokens.input"] == 10
-
-    def test_ignores_invalid_usage_quantities(self):
-        body = (
-            b'{"model":"claude-sonnet-4-6","usage":'
-            b'{"input_tokens":-1,"output_tokens":5,'
-            b'"cache_read_input_tokens":"50",'
-            b'"cache_creation_input_tokens":true}}'
-        )
-        result = extract_anthropic_messages_usage_from_json(body, None)
-        assert result == {
-            "model": "claude-sonnet-4-6",
-            "tokens.output": 5,
-        }
-
-    def test_handles_large_gzipped_body(self, headers):
-        """Body that decompresses past the legacy 64 KB cap should still parse.
-
-        Regression test for the silent 64 KB default in body_decoding.decompress_body
-        which used to truncate large model-provider non-SSE responses and cause
-        usage extraction to silently fail.
-        """
-        # Raw body > 64 KB (legacy STREAM_BUFFER_LIMIT) so the bug, if reintroduced,
-        # would truncate decompression output and break json.loads below.
-        big_text = "x" * (100 * 1024)
-        payload = json.dumps(
-            {
-                "id": "msg_1",
-                "model": "claude-sonnet-4-6",
-                "content": [{"type": "text", "text": big_text}],
-                "usage": {"input_tokens": 50, "output_tokens": 100},
-            }
-        ).encode()
-        compressed = gzip.compress(payload)
-        headers = headers(("Content-Encoding", "gzip"))
-        result = extract_anthropic_messages_usage_from_json(compressed, headers)
-        assert result is not None
-        assert result["tokens.input"] == 50
-        assert result["tokens.output"] == 100
-
-
-class TestExtractOpenAIResponsesUsageFromJson:
-    """Tests for OpenAI Responses API usage extraction."""
-
-    def test_extracts_model_tokens_and_cached_input(self):
-        body = json.dumps(
-            {
-                "id": "resp_123",
-                "model": "gpt-5.5",
-                "usage": {
-                    "input_tokens": 100,
-                    "output_tokens": 40,
-                    "input_tokens_details": {"cached_tokens": 25},
-                    "output_tokens_details": {"reasoning_tokens": 10},
-                },
-            }
-        ).encode()
-        result = extract_openai_responses_usage_from_json(body, None)
-        assert result is not None
-        assert result == {
-            "message_id": "resp_123",
-            "model": "gpt-5.5",
-            "tokens.input": 75,
-            "tokens.output": 40,
-            "tokens.cache_read": 25,
-        }
-        assert "reasoning_tokens" not in result
-
-    def test_missing_cached_input_details_does_not_emit_cache_read(self):
-        body = b'{"model":"gpt-5.4","usage":{"input_tokens":10,"output_tokens":5}}'
-        result = extract_openai_responses_usage_from_json(body, None)
-        assert result is not None
-        assert result == {
-            "model": "gpt-5.4",
-            "tokens.input": 10,
-            "tokens.output": 5,
-        }
-        assert "tokens.cache_read" not in result
-
-    def test_ignores_invalid_usage_quantities(self):
-        body = json.dumps(
-            {
-                "model": "gpt-5.5",
-                "usage": {
-                    "input_tokens": -1,
-                    "output_tokens": True,
-                    "input_tokens_details": {"cached_tokens": "25"},
-                },
-            }
-        ).encode()
-        assert extract_openai_responses_usage_from_json(body, None) is None
-
-    def test_invalid_cached_input_does_not_suppress_valid_input(self):
-        body = (
-            b'{"model":"gpt-5.5","usage":{"input_tokens":10,'
-            b'"input_tokens_details":{"cached_tokens":"bad"}}}'
-        )
-        result = extract_openai_responses_usage_from_json(body, None)
-        assert result is not None
-        assert result == {
-            "model": "gpt-5.5",
-            "tokens.input": 10,
-        }
-        assert "tokens.cache_read" not in result
-
-    def test_gzip_compressed(self, headers):
-        original = (
-            b'{"model":"gpt-5.3-codex","usage":{"input_tokens":42,'
-            b'"input_tokens_details":{"cached_tokens":7}}}'
-        )
-        compressed = gzip.compress(original)
-        headers = headers(("Content-Encoding", "gzip"))
-        result = extract_openai_responses_usage_from_json(compressed, headers)
-        assert result == {
-            "model": "gpt-5.3-codex",
-            "tokens.input": 35,
-            "tokens.cache_read": 7,
-        }
-
-    def test_truncated_gzip_stays_silent_but_diagnostic_returns_error(self, headers):
-        original = (
-            b'{"model":"gpt-5.3-codex","usage":{"input_tokens":42,'
-            b'"input_tokens_details":{"cached_tokens":7}}}'
-        )
-        truncated = gzip.compress(original)[:10]
-        headers = headers(("Content-Encoding", "gzip"))
-
-        assert extract_openai_responses_usage_from_json(truncated, headers) is None
-        usage, error = extract_openai_responses_usage_with_error_from_json(truncated, headers)
-        assert usage is None
-        assert error == "incomplete compressed body"
-
-    def test_cached_input_tokens_are_clamped_to_total_input(self):
-        body = (
-            b'{"model":"gpt-5.5","usage":{"input_tokens":5,'
-            b'"input_tokens_details":{"cached_tokens":7}}}'
-        )
-        result = extract_openai_responses_usage_from_json(body, None)
-        assert result == {
-            "model": "gpt-5.5",
-            "tokens.input": 0,
-            "tokens.cache_read": 5,
-        }
-
-    def test_extracts_usage_with_large_unselected_output(self):
-        body = json.dumps(
-            {
-                "id": "resp_large",
-                "model": "gpt-5.5",
-                "output": [
-                    {
-                        "content": [
-                            {
-                                "type": "output_text",
-                                "text": "x" * (100 * 1024),
-                            }
-                        ]
-                    }
-                ],
-                "usage": {
-                    "input_tokens": 20,
-                    "output_tokens": 9,
-                    "input_tokens_details": {"cached_tokens": 6},
-                },
-            }
-        ).encode()
-        result = extract_openai_responses_usage_from_json(body, None)
-        assert result == {
-            "message_id": "resp_large",
-            "model": "gpt-5.5",
-            "tokens.input": 14,
-            "tokens.output": 9,
-            "tokens.cache_read": 6,
-        }
-
-
-class TestStreamDecodeFeed:
-    """Direct tests for the bounded push-style streaming decoder."""
-
-    def test_gzip_happy_path(self, headers):
-        chunks: list[bytes] = []
-        parse = create_stream_decode_feed(headers(("Content-Encoding", "gzip")), chunks.append)
-        assert parse is not None
-        parse(gzip.compress(b"hello world"))
-        assert b"".join(chunks) == b"hello world"
-
-    def test_zstd_happy_path(self, headers):
-        chunks: list[bytes] = []
-        parse = create_stream_decode_feed(headers(("Content-Encoding", "zstd")), chunks.append)
-        assert parse is not None
-        parse(zstandard.ZstdCompressor().compress(b"hello world"))
-        assert b"".join(chunks) == b"hello world"
-
-    def test_supported_encodings_across_small_chunks(self, headers):
-        plaintext = b'{"model":"claude-sonnet-4-6","usage":{"input_tokens":42}}'
-        compressed_by_encoding = {
-            "gzip": gzip.compress(plaintext),
-            "deflate": zlib.compress(plaintext),
-            "zstd": zstandard.ZstdCompressor().compress(plaintext),
-        }
-
-        for encoding, compressed in compressed_by_encoding.items():
-            chunks: list[bytes] = []
-            parse = create_stream_decode_feed(
-                headers(("Content-Encoding", encoding)), chunks.append
-            )
-            assert parse is not None
-            for idx in range(0, len(compressed), 3):
-                parse(compressed[idx : idx + 3])
-            assert b"".join(chunks) == plaintext, encoding
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    def test_concatenated_zlib_members_same_callback(self, headers, encoding):
-        plaintext = b'{"model":"claude-sonnet-4-6","usage":{"input_tokens":42}}'
-        if encoding == "gzip":
-            compressed = gzip.compress(b"") + gzip.compress(plaintext)
-        else:
-            compressed = zlib.compress(b"") + zlib.compress(plaintext)
-        chunks: list[bytes] = []
-        parse = create_stream_decode_feed(headers(("Content-Encoding", encoding)), chunks.append)
-        assert parse is not None
-
-        parse(compressed)
-
-        assert b"".join(chunks) == plaintext
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    def test_concatenated_zlib_members_across_callbacks(self, headers, encoding):
-        plaintext = b'{"model":"claude-sonnet-4-6","usage":{"input_tokens":42}}'
-        if encoding == "gzip":
-            empty_member = gzip.compress(b"")
-            payload_member = gzip.compress(plaintext)
-        else:
-            empty_member = zlib.compress(b"")
-            payload_member = zlib.compress(plaintext)
-        chunks: list[bytes] = []
-        parse = create_stream_decode_feed(headers(("Content-Encoding", encoding)), chunks.append)
-        assert parse is not None
-
-        parse(empty_member)
-        parse(payload_member)
-
-        assert b"".join(chunks) == plaintext
-
-    def test_no_encoding_feeds_original_chunks(self, headers):
-        chunks: list[bytes] = []
-        parse = create_stream_decode_feed(headers(), chunks.append)
-        assert parse is not None
-        parse(b"hello")
-        parse(b" world")
-        assert chunks == [b"hello", b" world"]
-
-    def test_identity_feeds_original_chunks(self, headers):
-        chunks: list[bytes] = []
-        parse = create_stream_decode_feed(headers(("Content-Encoding", "identity")), chunks.append)
-        assert parse is not None
-        parse(b"hello")
-        assert chunks == [b"hello"]
-
-    def test_gzip_high_ratio_output_is_chunked(self, headers):
-        plaintext = b"A" * (STREAM_DECODE_CHUNK_LIMIT * 3 + 123)
-        chunks: list[bytes] = []
-        parse = create_stream_decode_feed(headers(("Content-Encoding", "gzip")), chunks.append)
-        assert parse is not None
-
-        parse(gzip.compress(plaintext))
-
-        assert b"".join(chunks) == plaintext
-        assert len(chunks) > 1
-        assert max(len(chunk) for chunk in chunks) <= STREAM_DECODE_CHUNK_LIMIT
-
-    def test_zstd_high_ratio_output_is_chunked(self, headers):
-        plaintext = b"A" * (STREAM_DECODE_CHUNK_LIMIT * 3 + 123)
-        chunks: list[bytes] = []
-        parse = create_stream_decode_feed(headers(("Content-Encoding", "zstd")), chunks.append)
-        assert parse is not None
-
-        parse(zstandard.ZstdCompressor().compress(plaintext))
-
-        assert b"".join(chunks) == plaintext
-        assert len(chunks) > 1
-        assert max(len(chunk) for chunk in chunks) <= STREAM_DECODE_CHUNK_LIMIT
-
-    def test_zstd_streaming_uses_writer_instead_of_decompressobj(self, headers, monkeypatch):
-        real_decompressor = zstandard.ZstdDecompressor
-        stats = {"stream_writer": 0, "decompressobj": 0}
-
-        class CountingZstdDecompressor:
-            def __init__(self):
-                self._inner = real_decompressor()
-
-            def stream_writer(self, sink):
-                stats["stream_writer"] += 1
-                return self._inner.stream_writer(sink)
-
-            def decompressobj(self):
-                stats["decompressobj"] += 1
-                raise AssertionError("streaming usage decoder must not use decompressobj")
-
-        monkeypatch.setattr("body_decoding.zstandard.ZstdDecompressor", CountingZstdDecompressor)
-        chunks: list[bytes] = []
-        parse = create_stream_decode_feed(headers(("Content-Encoding", "zstd")), chunks.append)
-        assert parse is not None
-
-        parse(zstandard.ZstdCompressor().compress(b"hello world"))
-
-        assert b"".join(chunks) == b"hello world"
-        assert stats == {"stream_writer": 1, "decompressobj": 0}
-
-    def test_gzip_error_logs_once_and_short_circuits(self, headers, mitm_ctx):
-        chunks: list[bytes] = []
-        with mitm_ctx() as log:
-            parse = create_stream_decode_feed(headers(("Content-Encoding", "gzip")), chunks.append)
-            assert parse is not None
-            parse(b"not gzip at all")
-            parse(b"more garbage")
-            parse(b"even more")
-        assert log.debug.call_count == 1
-        msg = log.debug.call_args[0][0]
-        assert "Streaming decompression failed" in msg
-        assert "gzip" in msg
-        assert chunks == []
-
-    def test_brotli_unsafe_encoding_logs_once_and_does_not_feed(self, headers, mitm_ctx):
-        chunks: list[bytes] = []
-        with mitm_ctx() as log:
-            parse = create_stream_decode_feed(headers(("Content-Encoding", "br")), chunks.append)
-        assert parse is None
-        assert log.debug.call_count == 1
-        assert "Streaming decompression skipped" in log.debug.call_args[0][0]
-        assert "br" in log.debug.call_args[0][0]
-        assert chunks == []
-
-    def test_zstd_error_logs_once_and_short_circuits(self, headers, mitm_ctx):
-        chunks: list[bytes] = []
-        with mitm_ctx() as log:
-            parse = create_stream_decode_feed(headers(("Content-Encoding", "zstd")), chunks.append)
-            assert parse is not None
-            parse(b"not zstd at all")
-            parse(b"more garbage")
-        assert log.debug.call_count == 1
-        assert "zstd" in log.debug.call_args[0][0]
-        assert chunks == []
-
-    def test_error_without_ctx_log_does_not_raise(self, headers):
-        # No mitm_ctx patch — ctx.log is unavailable.  Guard must swallow.
-        chunks: list[bytes] = []
-        parse = create_stream_decode_feed(headers(("Content-Encoding", "gzip")), chunks.append)
-        assert parse is not None
-        parse(b"garbage")
-        parse(b"more garbage")
-        assert chunks == []
-
-    def test_unsupported_encoding_logs_once_and_does_not_feed(self, headers, mitm_ctx):
-        chunks: list[bytes] = []
-        with mitm_ctx() as log:
-            parse = create_stream_decode_feed(
-                headers(("Content-Encoding", "compress")), chunks.append
-            )
-        assert parse is None
-        assert log.debug.call_count == 1
-        assert "unsupported content encoding" in log.debug.call_args[0][0]
-        assert chunks == []
-
-    def test_short_circuit_skips_decomp_fn_after_failure(self, headers, mitm_ctx, monkeypatch):
-        # Verify the broken flag actually prevents subsequent decoder calls.
-        # ``zlib.Decompress``
-        # is a C type whose ``decompress`` attribute is read-only, so we wrap
-        # the factory's return value in a proxy that counts delegations.
-        real_factory = zlib.decompressobj
-
-        class CountingProxy:
-            def __init__(self, real):
-                self._real = real
-                self.count = 0
-
-            def decompress(self, chunk, *a, **kw):
-                self.count += 1
-                return self._real.decompress(chunk, *a, **kw)
-
-            @property
-            def unconsumed_tail(self):
-                return self._real.unconsumed_tail
-
-            @property
-            def eof(self):
-                return self._real.eof
-
-            @property
-            def unused_data(self):
-                return self._real.unused_data
-
-        proxies: list[CountingProxy] = []
-
-        def factory(*args, **kwargs):
-            proxy = CountingProxy(real_factory(*args, **kwargs))
-            proxies.append(proxy)
-            return proxy
-
-        monkeypatch.setattr("body_decoding.zlib.decompressobj", factory)
-        chunks: list[bytes] = []
-        with mitm_ctx():
-            parse = create_stream_decode_feed(headers(("Content-Encoding", "gzip")), chunks.append)
-            assert parse is not None
-            parse(b"not gzip")
-            parse(b"more garbage")
-            parse(b"and more")
-        # Only the first chunk reaches zlib; later ones are short-circuited.
-        assert len(proxies) == 1
-        assert proxies[0].count == 1
-        assert chunks == []
-
-
-class TestDecompressBody:
-    """Direct tests for ``decompress_body`` — the non-streaming one-shot
-    path used by ``extract_anthropic_messages_usage_from_json`` and ``log_connector_usage``
-    to decompress full response bodies (up to
-    ``LARGE_RESPONSE_DECOMPRESS_LIMIT``) for JSON parsing.
-
-    Focus: verify the documented ``max_output`` cap is enforced during
-    decompression (not only via after-the-fact slicing) for codecs with
-    hard output-limit APIs.  Brotli's Python binding is best-effort; the
-    high-compression capture regression is covered in ``TestDecompression``.
-    """
-
-    def test_gzip_respects_max_output(self, headers):
-        # Regression: gzip path uses ``decompressobj.decompress(data,
-        # max_length=max_output)`` so zlib stops decoding at the cap
-        # rather than producing unbounded output.
-        plaintext = b"A" * (10 * 1024 * 1024)  # 10 MB, high compression ratio
-        compressed = gzip.compress(plaintext)
-        hdrs = headers(("Content-Encoding", "gzip"))
-        result = decompress_body(compressed, hdrs, max_output=64 * 1024)
-        assert len(result) <= 64 * 1024
-        assert result == plaintext[: len(result)]
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    def test_concatenated_zlib_members_after_empty_prefix(self, headers, encoding):
-        plaintext = b'{"ok":true}'
-        if encoding == "gzip":
-            compressed = gzip.compress(b"") + gzip.compress(plaintext)
-        else:
-            compressed = zlib.compress(b"") + zlib.compress(plaintext)
-
-        hdrs = headers(("Content-Encoding", encoding))
-        result = decompress_body(compressed, hdrs, max_output=64 * 1024)
-
-        assert result == plaintext
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    def test_concatenated_zlib_members_share_max_output_cap(self, headers, encoding):
-        first = b"A" * 8
-        second = b"B" * 8
-        if encoding == "gzip":
-            compressed = gzip.compress(first) + gzip.compress(second)
-        else:
-            compressed = zlib.compress(first) + zlib.compress(second)
-
-        hdrs = headers(("Content-Encoding", encoding))
-        result = decompress_body(compressed, hdrs, max_output=12)
-
-        assert result == first + second[:4]
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    def test_concatenated_zlib_empty_member_before_garbage_returns_empty(self, headers, encoding):
-        if encoding == "gzip":
-            compressed = gzip.compress(b"") + b"garbage"
-        else:
-            compressed = zlib.compress(b"") + b"garbage"
-
-        hdrs = headers(("Content-Encoding", encoding))
-        result = decompress_body(compressed, hdrs, max_output=64 * 1024)
-
-        assert result == b""
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    def test_concatenated_zlib_member_before_garbage_returns_decoded_prefix(
-        self, headers, encoding
-    ):
-        if encoding == "gzip":
-            compressed = gzip.compress(b"prefix") + b"garbage"
-        else:
-            compressed = zlib.compress(b"prefix") + b"garbage"
-
-        hdrs = headers(("Content-Encoding", encoding))
-        result = decompress_body(compressed, hdrs, max_output=64 * 1024)
-
-        assert result == b"prefix"
-
-    @pytest.mark.parametrize("encoding", ["gzip", "deflate"])
-    def test_invalid_zlib_first_member_returns_original_data(self, headers, encoding):
-        if encoding == "gzip":
-            corrupted = bytearray(gzip.compress(b"payload"))
-        else:
-            corrupted = bytearray(zlib.compress(b"payload"))
-        corrupted[-1] ^= 0xFF
-        compressed = bytes(corrupted)
-
-        hdrs = headers(("Content-Encoding", encoding))
-        result = decompress_body(compressed, hdrs, max_output=64 * 1024)
-
-        assert result == compressed
-
-    def test_zstd_respects_max_output(self, headers):
-        # Bug #10128: before the fix the zstd branch used
-        # ``decompressobj.decompress(data)`` which fully materialised
-        # the plaintext before slicing — defeating the bomb cap.
-        plaintext = b"A" * (10 * 1024 * 1024)  # 10 MB, high ratio → small payload
-        compressed = zstandard.ZstdCompressor().compress(plaintext)
-        assert len(compressed) < len(plaintext) // 100  # sanity: real high ratio
-        hdrs = headers(("Content-Encoding", "zstd"))
-        result = decompress_body(compressed, hdrs, max_output=64 * 1024)
-        assert len(result) <= 64 * 1024
-        assert result == plaintext[: len(result)]
-
-    def test_zstd_short_payload_returns_full_body(self, headers):
-        # When decompressed size is under the cap, return all of it.
-        plaintext = b"hello world"
-        compressed = zstandard.ZstdCompressor().compress(plaintext)
-        hdrs = headers(("Content-Encoding", "zstd"))
-        result = decompress_body(compressed, hdrs, max_output=64 * 1024)
-        assert result == plaintext
-
-    def test_brotli_large_input_caps_adaptive_chunk_size(self, headers, monkeypatch):
-        plaintext = _pseudo_random_ascii(STREAM_BUFFER_LIMIT * 3)
-        compressed = brotli.compress(plaintext)
-        assert len(compressed) > 64 * 1024
-
-        stats = _track_brotli_decompressor(monkeypatch)
-
-        hdrs = headers(("Content-Encoding", "br"))
-        result = decompress_body(compressed, hdrs, max_output=STREAM_BUFFER_LIMIT)
-
-        assert result == plaintext[:STREAM_BUFFER_LIMIT]
-        assert stats["max_input"] == 1024
-
-    def test_zstd_corrupted_returns_original_data(self, headers, mitm_ctx):
-        # Malformed payload should fall through to the outer
-        # ``except zstandard.ZstdError`` and return ``data`` unchanged,
-        # matching the existing gzip/brotli error contract.
-        hdrs = headers(("Content-Encoding", "zstd"))
-        garbage = b"this is not a zstd frame"
-        with mitm_ctx():
-            result = decompress_body(garbage, hdrs, max_output=64 * 1024)
-        assert result == garbage
-
-    def test_identity_returns_data_unchanged(self, headers):
-        data = b'{"hello":"world"}'
-        assert decompress_body(data, headers(("Content-Encoding", "identity"))) == data
-        assert decompress_body(data, headers()) == data
-
-    def test_gzip_empty_body_returns_empty(self, headers):
-        # Bug #10287: a valid gzip frame that decompresses to b"" must not be
-        # reported back as the compressed bytes.  Before the fix,
-        # ``return result if result else data`` on the success path handed the
-        # raw ~20 B framing to the caller, which then base64-encoded it into
-        # the network log.
-        compressed = gzip.compress(b"")
-        hdrs = headers(("Content-Encoding", "gzip"))
-        assert decompress_body(compressed, hdrs, max_output=64 * 1024) == b""
-
-    def test_deflate_empty_body_returns_empty(self, headers):
-        # Bug #10287: deflate shares the gzip branch but uses a different
-        # ``wbits`` — guard that the empty-body behaviour matches.
-        compressed = zlib.compress(b"")
-        hdrs = headers(("Content-Encoding", "deflate"))
-        assert decompress_body(compressed, hdrs, max_output=64 * 1024) == b""
-
-    def test_brotli_empty_body_returns_empty(self, headers):
-        # Bug #10287: same pattern as gzip for the brotli branch.
-        compressed = brotli.compress(b"")
-        hdrs = headers(("Content-Encoding", "br"))
-        assert decompress_body(compressed, hdrs, max_output=64 * 1024) == b""
-
-    def test_zstd_empty_body_returns_empty(self, headers):
-        # Bug #10287: same pattern as gzip for the zstd branch.
-        compressed = zstandard.ZstdCompressor().compress(b"")
-        hdrs = headers(("Content-Encoding", "zstd"))
-        assert decompress_body(compressed, hdrs, max_output=64 * 1024) == b""
